refactor(module-api): change names of modules
